+++ packages/api-jwt/src/index.ts
@@ -3,10 +3,10 @@ import 'tslib';
 import type { Client } from './lib/client';
 import type { ClientOptions } from './lib/types';
 
-export { PluginRoute as JWTAccessRoute } from './lib/routes/authorize/access';
-export { PluginRoute as JWTRefreshRoute } from './lib/routes/authorize/refresh';
-export { PluginRoute as JWTRevokeRoute } from './lib/routes/authorize/revoke';
-export { PluginMiddleware as JWTMiddleware } from './lib/middlewares/auth';
+export { PluginRoute as JWTAccessRoute } from './routes/authorize/access';
+export { PluginRoute as JWTRefreshRoute } from './routes/authorize/refresh';
+export { PluginRoute as JWTRevokeRoute } from './routes/authorize/revoke';
+export { PluginMiddleware as JWTMiddleware } from './middlewares/auth';
 export * from './lib/types';
 
 declare module '@sapphire/pieces' {
